helm_genie/requirements.txt
gitpython
langchain-community
langchain-openai
